pyproject.toml
[tool.poetry]
name = "liquidity-backtesting"
version = "0.1.0"
description = ""
authors = ["Jerzy <jerzybalcerzak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
demeter-fetch = "^1.1.10"
pandas = "^2.2.3"
zelos-demeter = "^1.0.1"
jupyter = "^1.1.1"
matplotlib = "^3.10.1"
plotly = "^6.0.0"
pydantic = "^2.10.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
